+++ src/view/ThingViewInterface.java
@@ -2,10 +2,10 @@ package view;
 
 public interface ThingViewInterface {
 
-	public void setX(int x);
-	public int getX();
-	public void setY(int y);
-	public int getY();
+	public void setX(double x);
+	public double getX();
+	public void setY(double y);
+	public double getY();
 	public void setImage(String imageFileName);
 	public void draw();
 }
